Refactor Graph into its own Component
